Basic HomeView Layout and Bug Fixes
Added a basic layout to HomeView that doesn't do much yet. Also added ScrollViews to onboarding screens that allow the user to scroll to dismiss the keyboard should they decide to. Additionally, a bug was fixed where the keyboard was staying present after logging in, this was fixed with the View+HideKeyboard.swift file in the Extensions folder. Finally, started implementing a basic CustomSegmentedPicker for picking the priority of a Project, added a custom back button to SignUpView to avoid strange centering behavior, and fixed a bug that would sometimes prevent the correct view from being shown after re-sending a verification code. This final bug was fixed by removing the @MainActor attribute from LoginController, as it forced the code to use a Task to asynchronously assign values that trigger the presentation of ConfirmCodeView. This asynchronous assignment is no longer necessary.