Added Widget Extension with Functioning Widget
Created a widget extension that shows a chart which details the user's work from the last week. This widget uses project and session data that is now stored in the app group via FileManager. The Widget Extension is also part of a Keychain Access Group to allow the user to sign in via the Widget Extension without them having to do anything. This functionality was not possible with Sign In With Apple and Google, so these features were removed since the LiveActivity will need direct access to the Amplify backend for it to work properly. Better interaction with the Session and Project JSON files in FileManager will be added in the next commit.