Added PlaceholderData for Widgets and Added TotalHoursWorkedWidget
TotahHoursWorkedWidget is a new widget that shows the total amount of hours the user has worked on the app. Placeholders were also added to the widgets to make them look better in the user's widget gallery. Finally, the widgets no longer attempt to sign in before accessing user data. This change was made because the data used by the widget does not necessitate that level of security, and there seems to be a strange Amplify.AuthError getting thrown when attempting to sign the user in via the widget (sometimes an error is thrown that says the user does not exist).