Projects and Sessions are Modified when Signing in and out
Signing out will delete all the Projects and Sessions on the user's device, while signing in will get all of the user's projects and sessions from the Amplify backend and save them to the user's device storage. This is done to make the widget workflow smoother when signing in and out of the app.