Removed Live Activities
Live Activities have been removed from the app because the initial functionality that was intended does not appear to be possible at this time. The Live Activity was intended to show a stopwatch the user can use to always see how long a session has been running for. This works if pausing the stopwatch is not supported. That's because it does not appear to be possible to start and stop a Timer object in a Live Activity. At this time, I also cannot find any example of a Timer object working and updating the UI in a Live Activity. The only way to run a timer in a Live Activity appears to be by using a Text View. Although this works, it does not appear to be possible to store the value associated with this Text View, nor does it appear to be possible to pause the timer this way. Such functionality is necessary because the main app supports pausing the timer. This will be explored again at a later date.