App Now Tracks Timer Duration Despite Going to Background
Using the .onForeground and .onBackground modifiers, OngoingSessionController now stores the date at which a user puts the app in the background if a session is currently active. This allows the user to see an up-to-date timer value that includes the time for which the app was closed once they re-enter the app.