Added In-App Purchase Capability and New HomeView Dependency Injection
Altered the login flow so that the currently logged in user is provided to HomeView when it's initialized. This was done because HomeView needs access to the logged in user's new isPremium property so it can determine whether or not to offer the user an Upgrade to a premium account. The alternative to this is fetching the loggedInUser after  HomeView is initialized, which is not ideal since it's less predictable and clean. UpgradeView has been added and the In-App Purchase Capability has been added to the Xcode project, but no other StoreKit functionality has been added to work with the Premium item added in App Store Connect.